fastboot: TCP transport
For devices which do not have USB, we can run fastboot over the
network. This requires corresponding TCP support in the device's
Fastboot server, which is supposed to listen for connections on
port 1234. Fastboot client refactored in order to support the
transport specification option, which for TCP looks as follows:
   -T tcp:<device IP or HostName>

BUG=25967857
Change-Id: I5ef66527d44830d46c1ef09e7fa9dc7da7dab81c
Signed-off-by: Viorel Suman <viorel.suman@intel.com>
Reviewed-by: Beare, Bruce J <bruce.j.beare@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>