enable automatic bounds checking for init
This enables the bounds sanitizer for the init code, providing automatic
bounds checking for statically sized arrays.

The performance overhead is generally in the 5-10% ballpark for code
performing computations where they end up providing significant bounds
checking coverage. For processes like init, there's no significant cost
since it's completely bounded by the cost of system calls. It's usually
significantly cheaper than automatic integer overflow checking.

This feature currently lacks any coverage of dynamic allocations but
libc++ has support for setting a debug level to turn on features like
bounds checks. It could be tweaked to offer a lighter mode for hardening
(i.e. no logging, only abort).

It would make sense to enable the object-size sanitizer alongside this
to cover more than just arrays and for better coverage of arrays, but
some minor cases of undefined behavior would need to be fixed to take
advantage of it. The bounds and object-size sanitizers have a lot of
overlap but it doesn't result in unnecessary overhead as long as both
are set to handle errors in the same way, as the compiler will merge all
of the redundant checks.

Change-Id: Ic619fdb30590d9d33dbe793f95403497b25c6610