enable automatic bounds checking for init
This enables the bounds and object-size sanitizers for the init code,
providing automatic bounds checking for statically sized arrays and
anywhere that the object size intrinsic works. The sanitizers have a lot
of overlap (bounds is close to a subset of the object-size sanitizer
with optimization enabled), but when they're set to use the same
trapping function the redundancies are optimized out.

The performance overhead is generally in the 5-15% ballpark for code
performing computations where they end up providing significant bounds
checking coverage. For stuff like init, there's no significant cost
since it's completely bounded by the cost of system calls. They're
significantly cheaper than integer overflow checking in most cases.

The bounds sanitizer could essentially be enabled across the board if
the performance hit wasn't a concern while the object-size sanitizer
tends to run into previously undetected undefined behavior, although
it's fine here (and in libbinder, libcutils and libutils). It tends to
find (usually minor) issues not caught by Valgrind or ASan.

These lack any coverage of dynamic allocations, but libc++ has support
for setting a debug level to turn on features like bounds checks. It
could be tweaked to offer a lighter mode for hardening (i.e. no logging,
only abort).

Change-Id: Ic619fdb30590d9d33dbe793f95403497b25c6610