Update patch set 1
Patch Set 1:

this gets me the correct error message 100% of the time, even on dimitry's N7. the weird (to me) kernel behavior might explain why this is so much more frequent a problem on N7, if there's some variation by kernel?

the only real alternative i can think of is doing a poll(2) in the read/write loop in SendLargeFile and aborting and reading the error if there's something to read. but i'm not sure that would be sufficient either... i'm not sure exactly what the conditions are to cause the error message to be dropped. (it's weird too because the 8-byte read gets the FAIL<binary-43>, but then when we try to read the 43 byte error message, that read returns 0. i wonder if that wasn't two writes on the server side whether that would also solve the problem?)

Patch-set: 1