Update patch set 1
Patch Set 1:

> other projects we actually moved out of CrOS instead of forking. 
 > is there a reason we don't do the same for crash-reporter ?  seems
 > like we should be dropping it from platform2 and pointing the code
 > to here.  which means we don't want to drop CrOS code.

For one, when it was decided to put crash reporter into system/core, it made doing a cross-checkout more difficult, since I didn't think there was a way to just check out a specific subdirectory of a git repository, and we probably wouldn't want CrOS to check out all of system/core just to get crash_reporter and metrics.

Additionally, ideally I'm sort of against putting a ton of #IFDEFs into the code to build for the two different platforms when the changes include program logic changes, not just changes to constants/variables.  Between the changes needed to handle SELinux enforcement, differences in the filesystem layout, a few differences in Bionic vs libc, and platform differences like udev, init, system properties, and os-release.d, there would be a lot of IFDEFs that would make code management much less simple to read.  This would get even more difficult if/when we start to consider adding support for Java crashes.

Patch-set: 1