Update patch set 16
Patch Set 16:

I wonder how well this is going to respond to signals?  The binder thread is going to be blocked in an ioctl when the signal comes.  I suspect the Daemon installs some signal handlers on your behalf, which will shut down the main thread.  You should probably experiment a little and figure out what this binder code is going to do on a signal?

Patch-set: 16
Reviewer: Gerrit User 1057759 <1057759@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0