Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

There is a gTest coming; but it is part of the LOG_ID_SECURITY CL (blocked until logcat long-arguments patches all clear). I suspect I should have added a simple gTest for the condition you identified in the review as part of this CL (if it set persist.logd.security to "true" it would have caused a 'pass' result with an empty ro.device_owner (wrong). It is impossible to test ro.device_owner in a gTest due to the nature of ro.* properties though; they must be done manually with intervening reboots, and would not work once the SetupWizard does its job.

Patch-set: 1
Label: Code-Review=-1