Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

A small piece of cleanup that has an impact on a more refined determination of realtime and realtime. I have confirmed this fixes the bug where time is in flux during SetupWizard and also allows the -t and -T flag to work when they specify time when the realtime is in the 1970s (before time is set, before wifi connects to ntp, before timezone is set).

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=+1