Update patch set 1
Patch Set 1:

timestamp == first log will be a number 2 (very unlikely unless the log is yet to be pruned) or greater; but impossible to distinguish at this point if that is the first matching/found entry in the log, or the last entry in the log. Worth sleeping in either case.

As for timestamp < first log entry, it is less than 2 (0 is invalid, 1 means nothing was found).

Patch-set: 1