Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

I want to give this time to percolate, we are not in a rush to apply this. I may decide that there are other algorithmic improvements and optimizations in the logger that may offset the latency and noise issues.

eg:

1) The complaints about "chatty" were supposedly addressed by the "cap" on "chatty" as 10-fold larger than the logged volume. I am not sure if I have heard anyone complain since that change.

2) One of the sources of latency occurs in the later pruning steps, there is no watermark held per log buffer id, and thus pruning in the "main" buffer could be forced to iterate through the whole logs just to get to the initial "main" buffer logs to start pruning them. I expect adding this watermark may have a positive effect nullifying the need to elevate logd.writer priority.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1
Label: Verified=+1