fastboot: add TCP protocol.
This CL implements a TCP protocol for use with fastboot. Protocol
description is given in fastboot_protocol.txt, some examples of
expected behavior can also be found in tcp_test.cpp.

This CL also adds a SocketMock class for use in tests. This provides a
way to mock out network read/write behavior and ensure that the TCP
protocol reacts appropriately.

There is also a small fix to socket_test.cpp to try a range of ports in
case of random port conflict, which I just ran into.

Bug: http://b/26558551
Change-Id: If53a514a534489c617db32c4fea8819949121282