Update patch set 4
Patch Set 4:

A few concerns:

1) Logging should be conditional based on the value of persist.logd.security and ro.device_owner. This is an unconditional log, so we're logging EVERY adb push/pull operation on EVERY device.

Also, there's no guarantee that these logs will only ever be consumable by the device owner. There's going to be a big desire by lots of people to get these logs from the security log. These logs need to be encrypted to ensure that only the device owner can access them, and a leakage of the logs to other parts of Android can't be used to compromise the user's activities on the device.

Patch-set: 4