Update patch set 2
Patch Set 2:

(1 comment)

Sorry about not reviewing this for so long, it fell off of my radar (and testing windows stuff is a massive pain with our VMs; they're abysmally slow, presumably because of all of the security crapware on them).

I think I found out the reason why the output gets mangled so badly on windows. The behavior of the windows console seems to differ from ANSI terminals when reaching the end of a line. If you have a 10 column wide terminal, and echo "012345678\bx9\byfoo", you get the following results:

Linux:
  01234567xy
  foo

Windows:
  01234567x9
  yfoo

This seems to be controlled by ENABLE_WRAP_AT_EOL_OUTPUT, but the two behaviors to choose from seem to be "immediately wrap" and "write to the last character of the line". I don't know if there's a sensible way to fix this, short of having a full terminal emulation library.

Patch-set: 2
Label: Verified=0