Update patch set 3
Patch Set 3: Code-Review+2 Verified+1 Presubmit-Ready+1

(1 comment)

> Maybe you guys could consider using a physical Windows machine if
 > you're having problems with VMs? I mean, Windows probably has the
 > most market share of adb users.

That's something worth considering, but it might be a worse use of time than just spending the hour or two it'll probably take to properly fix the VM. I've got mine in a reasonably workable state now (not quite tolerable for Windows compiler toolchain testing, but definitely good enough to iterate on adb).

 > I don't see a sensible way to fix this behavior. My opinion is to
 > leave this broken end-of-line behavior. My overall vision with the
 > various Windows console changes that I've done is to make things
 > more *comfortable* for Windows users (i.e. Ctrl-C, line-editing
 > keys, no double-echoing, SIGWINCH, etc.), but don't do a full-blown
 > terminal since someone who needs that level of support can run
 > their own sshd and login with putty or other 'real' terminal
 > emulator.

Yeah, I was hoping that I had missed something, and there was a one line fix for everything. Barring that, this seems reasonable to me.

One thing that might work is to break things in the other direction. If we lie, and tell adbd that the terminal is 1 column narrower than it actually is, I think it might fix more things than it breaks. Programs that print however many characters are in a row, and rely on solely on text wrapping without using explicit newlines will break, but that seems unlikely?

Patch-set: 3
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Presubmit-Ready=0
Label: Verified=+1