Update patch set 1
Patch Set 1: Code-Review+1

Requires Windows Vista, but that's ok because build/core was incremented by Hines a few weeks ago.

General comment: errno is not set in these wrappers. I'm guessing that you're going to deal with this later. You can see the headaches that sysdeps does with this.

Patch-set: 1
Label: Code-Review=+1