Update patch set 1
Patch Set 1:

(3 comments)

I would like to see something that got truncated end up matching. We need to embody the 'normal' truncation that occurs within the find algorithm when matching, I know this code does not do well with that yet.

Also, the length PROP_NAME_MAX might be a bit *long*, we prepend a minimum of log.tag.<tag>, and a maximum or persist.log.tag.<tag>, maybe we should use PROP_NAME_MAX - 16 (size of "persist.log.tag.") as a universal truncation so that we get consistency in matching?
strlcpy(last_tag, tag, sizeof(last_tag));
I had chosen dup (now that I remember) because it made no assumptions about truncation, so this is an API change in some respects. property check however deals with a match to a truncated entry, so this is better than the code I have here right now at consistency.

Patch-set: 1