Update patch set 3
Patch Set 3: Code-Review+1 Verified+1

Re-added the memory order and made it better. I like that memory_order_consume is lower priority, but the system call to close probably did all the heavy lifting for that one and so it is merely a preference. It affects the compiler optimizer whereas memory_order_relaxed does not.

Patch-set: 3
Label: Verified=+1