Update patch set 3
Patch Set 3: Code-Review+1 Verified+1

Did not have to explicitly state memory_order_relaxed, but in the face of the other atomic calls, adds clarity of purpose. By not glomming the error messages together, we handedly deal with the chance we will get an old cached value and accept the consequences when we print.

Patch-set: 3
Label: Code-Review=+1
Label: Verified=+1