Update patch set 4
Patch Set 4: Code-Review+1 Verified+1 Presubmit-Ready+1

Turned on atomics in a limited sense here, to allow an error-path to sniff for a possible value. Reading the atomic is done directly (without the atomic handler) when in locked context, is this OK boehm@, or do I have to use the handlers in all cases?

Patch-set: 4
Label: Code-Review=+1
Label: Presubmit-Ready=0
Label: Verified=+1