Update patch set 5
Patch Set 5:

(1 comment)

Nice, if this works it's a great simplification!

If we could get socket_test.cpp compiling and passing on Windows, I think that would add a lot of confidence that adb_socketpair() works the same on all platforms, but there seems to be a fair bit of Unix-specific threading code in there at the moment so I'm not sure how painful that would be.

Patch-set: 5