Update patch set 1
Patch Set 1:

Probably the wrong place for this, but difficult to suggest an alternative given that there can only be _one_ auditd. logd has been a passive up to this point, and can be disabled or removed from the platform. There is a developer setting to turn off logging. logd.auditd can be disabled by the platform.

Patch-set: 1