Update patch set 1
Patch Set 1:

in system/core/logd/main.cpp:

    bool auditd = property_get_bool("logd.auditd",
                                    BOOL_DEFAULT_TRUE |
                                    BOOL_DEFAULT_FLAG_PERSIST);

May need to get special treatment to _prevent_ actually disabling logd.auditd thread in this mode, and instead just disabling actually writing them to the associated destinations (events, main and klog(?)). For clarity, when user selects that logs are disabled, it blocks content going to events and main, but not klog, and onDataAvailable continues to be called so it does _not_ prevent your current path from functioning.

Patch-set: 1