Update patch set 1
Patch Set 1: Code-Review-1

The underlying "problem" is two instances of this pattern in adf_find_nodes():

    if (some_condition) {
        ret = -errno;

clang's static analyzer objects to the case where this branch is taken and ret >= 0.  This assumes we go into the branch with errno <= 0, which isn't actually possible.

Assuming abort() is the best way to educate clang, I would change this to:

    if (some_condition) {
        ret = -errno;
        if (ret >= 0)
                abort();

Patch-set: 1