Improve adb's LinePrinter for dumb terminals.
I started down this path, but I'm not sure it's a good idea. What should
"dumb" mean? We can continue like this (by overprinting with spaces), but
is that really the intent for "dumb"? Or should we be more like the old
sync output? Or should we just be mostly silent? Or like the old output
but with the overall percentage (but not the per-file percentages)?

Change-Id: I1d7fbbb8b01b9b9035e703fd5aa00e3d980911ec