Update patch set 2
Patch Set 2: Verified+1

This at least makes it possible to do so (add packages healthd and libbatteryservice).  

At least 2 of our starter boards already try to run it in their Android-derived init scripts).  The flash delta is tiny but the RAM USS is about 365K, not huge but not insignificant, so I wasn't going to add it to the base product unless there's a consensus to do so.

Patch-set: 2
Label: Verified=+1