Update patch set 2
Patch Set 2:

Everyone is agreed that using !mBatteryDevicePresent for triggering fake battery logic is the better way to go.  If you want to send a second patch that takes that approach then I think we're set to go.  But we don't want to add more conditions around the existing code, which isn't doing what is wanted for your situation.

Patch-set: 2
Label: Verified=0