Update patch set 2
Patch Set 2: Code-Review+2

Is the default for libminijail to have an unrestricted capability bounding set (eg, every capability is allowed?). If so, you may want to modify libminijail so that the default bounding set is 0, and require the user to explicitly add capabilities to the bounding set. That way, you have the tightest, safest defaults, and people need to go out of their way to relax the capability bounding set.

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2