Update patch set 2
Patch Set 2:

What we need is a way for the list to include/not include this based on whether the device is supposed to have a webview or not. (currently the only devices that are supposed to lack webview is Android Wear). mips64 devices *should* have a webview in general, and while we don't currently provide a prebuilt, anyone building a mips64 device should have built webview for it and included it in the platform build, so this library should also be present.

Patch-set: 2