Update patch set 1
Patch Set 1:

> @stephen, do you want to submit this with the llvm rebase for the
 > same topic? I can take care of the change in simpleperf?

Can you look at the simpleperf differences (mostly because of the iterator changes)? You should only need the changes that Pirama uploaded for external/llvm, clang, and compiler-rt to "mma" it. There are a lot of other projects that we have ongoing to get the rest of the build in place.

Patch-set: 1