Update patch set 7
Patch Set 7:

Hi Rom,

Thanks for pointing me in the right direction. We do rely on the fact that "property_set" for a "ro.boot.abc" property can be called only once, so the one from DT will stick since "process_kernel_dt()" is called before "process_kernel_cmdline()". Thus the valid option for DT case is to set the content of "/proc/device-tree/firmware/android/console" file to be "tty0:ttyS0" if tty0 and ttyS0 are to be supported. A similar aggregation functionality like in kernel parameters case IMHO is not possible in DT case since this implies the presence of multiple files, all of them having the same, "console", name in "/proc/device-tree/firmware/android" directory.

Patch-set: 7