Update patch set 1
Patch Set 1:

I didn't want to use a thread for this because I wanted to keep debuggerd single-threaded so that there's no danger of a fork happening while the signal sender is in the middle of doing something dangerous (does ALOGE allocate?)

This leads to there being 4 debuggerd processes running at all times, though. I'm thinking of maybe using clone with only CLONE_THREAD to hide the process (and also simplify the cleanup stuff when we decide to tear everything down), but that might be too clever.

Patch-set: 1
Label: Verified=0