Update patch set 1
Patch Set 1: Code-Review+1 Verified+1

Discovered a regression in this test if we increase the tag length beyond 4068 bytes (as part of https://android-review.googlesource.com/#/c/208045). Opted here to try to print as much of the information as possible instead of truncating to <100 (before the referenced CL) dropping it on the floor (what happens after the referenced CL). This is not advocacy for malformed messages, but a DYB effort to provide as much usefulness and functionality to the log printer.

Patch-set: 1
Label: Code-Review=+1
Label: Verified=+1