adb: use a custom thread to poll for usb devices on mac.
On mac, if the adb server kicks a transport on some error, mac usb driver
will not report a new usb device. So instead of relying on mac usb driver
to report new usb devices, this CL uses a loop to search for usb devices
not exist before. Note that this is also the behavior on windows and linux
host.

Add `adb reconnect` command to test the reconnection ability of adb server.

Bug: 25935458
Change-Id: I890e0eb1fae173f2e7a0c962ededa294d821e015