adb: monitor status of transports and kick invalid transports.
If transport on device is kicked, the device will skip all
packets except A_CNXN packet. However, the host may not detect
that and send A_OPEN packets as it is connected. This problem
can be reproduces by running `adb reconnect device` for a few
times.

the solution in this patch is to monitor transport activities,
and if there is no A_OKAY packet replied for an A_OPEN packet
for 1 second, the adb server will kick the probably invalid
transport. It also kicks transports in offline state for >=3
seconds. To verify the solution, run the following commands:
adb reconnect device
adb shell # if it fails, the problem described above happens
adb shell # it should success as the adb server has kicked and reconnected the transport.

Bug: 24370690
Change-Id: Ie8edb54ef5a1a6bdee3244739dfb7e53ee9b1718