adb: use heartbeat packets to verify transports.
If the transport on device is kicked, the device will skip all
packets except A_CNXN packet. However, the host may not detect
that and send A_OPEN packets as it is connected. This problem
can be reproduces by running `adb reconnect device` for a few
times.
The solution in this patch is to let adb host server send a
heartbeat A_OPEN packet to each device regularly. If a device
is in connected state, it should reply a heartbeat A_CLSE
packet. If the adb host server doesn't receive a heartbeat
A_CLSE packet from a device for preset times, it will kick the
probably invalid transport for the device.
It also kicks transports staying in kCsOffline state for a period too long.
To verify the solution, run the following commands:
adb reconnect device
sleep 1
adb shell # it should success as the adb server has kicked and reconnected the transport.

Note that the heartbeat verification method doesn't need change
on device, so it can work with devices without this change.

Bug: 25935458
Change-Id: Ie8edb54ef5a1a6bdee3244739dfb7e53ee9b1718