Update patch set 4
Patch Set 4:

(1 comment)

It doesn't change battery life on device side. the adbd on device will sleep and wait for new actions when idle. But on host side, the adb host server doesn't stop when devices are disconnected from USB. As you said, there is a thread looking for new devices, a thread looking for new emulators, and now a new thread kicking invalid transports. But I think this patch isn't terrible enough to affect the whole battery life of the host, no?

Patch-set: 4
Label: Verified=0