init: drop use of AID_READPROC and hard-code
Change init to not use android_filesystem_config.h. This
removes the dependency on this interface.

The choice to hardcode the readproc uid value is twofold:
Firstly hardcoding "readproc" or "3009" always results in
a constant being embedded. Secondly, converting the uid_t
from a call to getpwnam() to a string to use in the mount()
call for procfs was more trouble than it's worth, that
code should be dead simple.

Change-Id: I2b383b04cfb421a5f9f32ae2a92b20b3b75cf4a9
Signed-off-by: William Roberts <william.c.roberts@intel.com>