Update patch set 2
Patch Set 2:

(3 comments)

The clearest advantage I see to this are outlined in b/27999086.

What that bug fails to point out (thus far) is android_ids is a _static_ and shows up everywhere in the various binaries and is not deduped in executables.

One will find _two_ copies of this in init, one for libc_static, and another for init/util.o.

Patch-set: 2