Update patch set 1
Patch Set 1:

(1 comment)

Of course, none of this is worthy of the goals of the current code change; and require the future changes in bionic getpwnam to absorb the libpackageparser and partner adjustable content (fixed in /system/etc/passwd) mechanics. This is just a standalone incremental first pass in Phase 3 of the bug.

Patch-set: 1