Update patch set 2
Patch Set 2: Code-Review-1

I am concerned we have partners using this as well, cleansing this may require a 'generation' and it may be too soon to deprecate the,.

Also, I do not want to populate /system/etc/passwd with these immutable ids, it give the impression that they can be changed, and they can not. I am not tied to that though. Hiding them from view will make it harder (not impossible) to construct them if we wanted to add them to the passwd and groups files; the host tool that does so would have a dependency on this bionic file ...

Retaining them in a _private_ header file might be a necessary evil to permit the host tool to be constructed.

Patch-set: 2
Label: Code-Review=-1