Update patch set 2
Patch Set 2:

> > I am concerned we have partners using this as well, cleansing
 > this
 > > may require a 'generation' and it may be too soon to deprecate
 > > the,.
 > 
 > not at all. we need to set an example, and moving Nexus devices off
 > this is the best way to do that.
 > 
 > > Also, I do not want to populate /system/etc/passwd with these
 > > immutable ids, it give the impression that they can be changed,
 > and
 > > they can not.
 > 
 > this part i agree with. i don't see any particular advantage to
 > moving everything to files, and i definitely worry about it
 > encouraging OEMs to accidentally/deliberately break stuff.

Consistency of interface and design. We should make it hard for them
to accidentally break it, if they intentionally break it, it should
fail CTS.

We did this in selinux, and it works well. What if we make a CTS test
to pull the /etc/passwd /etc/group files and ensure that they did not
modify the base aosp set. At build time, we can bundle the AOSP generated
passwd/group file inside of CTS and check. It works pretty well, see the
security tests we added in CTS for selinux. Are their sanity tests on AID
today in CTS, that 1000 is really "system", etc?

The tools in fs_config are the interfaceto generating these files. The only
was you should be able to generate a bad config is by modifying the tools.

The only way you should be able to break it is by:
disabling verity, remounting system, and writing the file directly. Which
will not be supported, and a big fat do not modify header at the top of
those files makes it clear (as well as the ro file system)

I have the tool (see my patches on build) that can generate the passwd
file based on a base config provided by AOSP and OEM supplied files. It
prevents OEMs from defining an AID not in their range, as well as duplicates.

It also prevents the base config from generating AIDs in the app range,
which is a check we never had before, which might help someone who is hacking
on AIDs realize that's a no-no.

I think moving to what I am describing is more extensible, easier to maintain,
and the CTS checks give us greater assurance on checks. Just my 2 cents though.

 > 
 > > Retaining them in a _private_ header file might be a necessary
 > evil
 > > to permit the host tool to be constructed.
 > 
 > there's no such thing as private in the platform. you yourself are
 > regularly responsibly for adding #include <private/...> all over
 > the codebase. if we mean to clean this up -- and i think we do --
 > we need to just do it.

Patch-set: 2
Label: Verified=0