Update patch set 4
Patch Set 4:

(1 comment)

Normally I like to go in when I make changes, and improve the code. Today is an exception.

Agreed. There are a few interfaces inside logd where I made the design mistake of const char * passed back to a new owner with the responsibility of freeing (predates C++11 acceptance into Android); There is a pending 'cleanup' to switch to std::string (or std:unique_ptr, whatever). I expect that to be outside the scope and purpose of this CL which is to allow us to immediately distance ourselves from android_ids[].

There are 2 ToDo items:
1) Once we have mechanics for libc+libpackageparser, move getpwuid() call into logd.daemon for serialization, MAC and DAC. Deprecate libpackageparser code, but keep libpackageparser library. Not doing this today since getpwuid() is MUCH more efficient in current thread than rpc'd to logd.daemon; but a cost we are willing to pay for maintainability at that juncture.
2) See what we have then (or vice versa), and perform std::string cleanup at our leisure ;-}

Patch-set: 4