Change wait_value_ to an atomic type.
There is a race condition that can occur in the Wait() and Wake() calls.
Thread A has called the Wake() function and completed it.
Thread B comes along after Thread A has completed, and is in the Wait()
call. It grabs the mutex and checks wait_value_, but this thread does
not see the value has changed. It calls pthread_cond_timedwait(), but
the Wake() has completed so no pthread_cond_signal call will ever happen.
Now the only way to continue is to wait for the timeout.

The fix is to make wait_value_ an atomic value to guarantee that all threads
will see the same value.

Bug: 28184010
Bug: https://code.google.com/p/android/issues/detail?id=206856
Change-Id: I08ada2f9bf074e177b7e7ddfd271416241833f62