Update patch set 1
Patch Set 1:

Can you explain more about how this race condition occurs? There's only one mknod in all of ueventd, and it's properly surrounded by calls to setfscreatecon(), which guarantees that the device will be labeled atomically.

This is either a kernel bug, or this device node is getting created outside of this code path, both of which should be fixed and not worked around.

Do you have steps to reproduce this? This isn't something I've seen on other devices.

Patch-set: 1
Label: Verified=0