Update patch set 3
Patch Set 3:

A treatise.

b/25792367 #53, #54, #55. Basically init has a performance issue. For one (out of many) it uses a UNIX domain socket to be told about updates to the properties and other system control activities, and this can be overloaded, and _is_ overloaded by the test 1/30 tries. In addition, properties are in shared memory and lockless, suffering somewhat from memory synchronization issues in the extreme, but work fine for occasional updates. properties are _not_ designed for, or should ever be used for, rpc. Their adjustments should be few, and in the context of this test managed at a controlled and reasonable pace. The 20ms is still _not_ reasonable, but if we add any much more delay than go/apct will time out on the test and report a failure. 20 ms is the best compromise to allow the test to pass based on the entire potpourri of logic coverage.

We are _not_going to fix properties, it is what it is, and this test should not be used to also test the limits of Android property management. The 1/30 failures reported by go/apct is a ghost corner case.

Patch-set: 3