Update patch set 4
Patch Set 4:

This seems to be weakly implicated in b/28211927.  Reviews appreciated.

I'm still thinking about possible stress tests. I haven't had a lot of luck with stress tests for memory model issues in general.

The only test I could find for any of this is StrongPointer_test.cpp, which is essentially a smoke test for the simplest part of the functionality in a single-threaded setting.

Patch-set: 4
Label: Verified=0