Update patch set 6
Patch Set 6:

(2 comments)

The really old version of the code, before conversion to standard atomics inside, had a full fence before the atomic RMW, but not after, which effectively gave it release, but not acquire semantics. I believe they were always documented as release operations.  This was a bit stronger in that it had an undocumented side effect of ordering other unrelated operations.  But it did not help here.  Read accesses in the destructor could still be reordered with the final atomic decrement, since there was no barrier separating them.

I believe the code we now generate is the same for 32 bit, but we should be generating a release exclusive store instead of the fence on ARMv8. Both have the same issue.

Patch-set: 6
Label: Verified=0