Update patch set 1
Patch Set 1:

i'm assuming this isn't an option (like it is for modprobe(1)) because the idea is that they'll want to ship a new kernel without having to update the modules *or* the init.rc files? might be worth explicitly mentioning that in the checkin comment.

otherwise fine by me unless nnk has concerns?

Patch-set: 1