Update patch set 1
Patch Set 1: Code-Review+1

I had the same question as Elliott. Perhaps we should make this an argument in the init.*.rc file, so anyone who writes this could choose to add this, or not add it.

Alternatively, perhaps we could introduce a new builtin type:

  insmod_no_vermagic /path/to/module

which would be the same as insmod, but without the VERMAGIC check.

Regardless, this is fine with me.

Patch-set: 1