Update patch set 3
Patch Set 3:

I ended up changing dealloc to return void, since no clients look at the result anyway.  And I added a comment that SharedBuffer is deprecated and should not be used for new code, thus weaseling out of the task of trying to document what it does and what the thread-safety rules actually are.

Patch-set: 3