Update patch set 1
Patch Set 1:

It will break many tools based on perf, even though they can work around it by excluding kernel code. The long-term intention is to migrate this to perf_event_paranoid=3 when it's supported to *completely* disable perf events but the progress on it landing upstream is slow. It's a grsecurity feature (PERF_HARDEN) but it was adopted by Debian for their vanilla kernel too, and is enabled by default.

Android could carry a similar patch as Debian out-of-tree until the feature can be cherry-picked from mainline, but I'm not sure if that would be approved of.

Patch-set: 1