Update patch set 2
Patch Set 2:

> (1 comment)

For your refererence: A7.9 and A7.10 (pg 206 and 207) of the Second Edition of The C Programming Language. Clearly indicates that 0 or 1 is the result, and the manpage for stdbool indicates that http://pubs.opengroup.org/onlinepubs/009695399/basedefs/stdbool.h.html true false will expand to the same thing.

The compiler does seem smart enough to not create jumps for this:

(the delta between approaches)

< 	cmpl	$1, -20(%rbp)
< 	sete	%al
---
> 	cmpl	$0, -20(%rbp)
> 	setne	%al

Patch-set: 2
Reviewer: Gerrit User 1013433 <1013433@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0