Update patch set 2
Patch Set 2:

The issue is that in AOSP /init does the parsing of the vars, in N logcat does the parsing of the vars. N has _less_ expansions available as well.

The errors I get are:

[   64.471944] ext4_utils: property 'persist.logd.logpersistd.buffer' doesn't exist while expanding '${persist.logd.logpersistd.buffer}'
[   64.471986] ext4_utils: setprop: cannot expand '${persist.logd.logpersistd.buffer}'
[  112.560539] ext4_utils: property 'persist.logd.logpersistd.buffer' doesn't exist while expanding '${persist.logd.logpersistd.buffer}'
[  112.560577] ext4_utils: setprop: cannot expand '${persist.logd.logpersistd.buffer}'
[  117.388272] ext4_utils: property 'persist.logd.logpersistd.buffer' doesn't exist while expanding '${persist.logd.logpersistd.buffer}'
[  117.388313] ext4_utils: setprop: cannot expand '${persist.logd.logpersistd.buffer}'

which occur when the properties are <empty> or non-existent. We do _not_ want the properties to propagate into the trampoline in these cases, so the behavior is desirable. The issue of course is that the parser does not know this, and it _should_ produce an error for all other use cases. Now, if the error (and the failure to exec) could be suppressed because the command is setprop, maybe; but we would also need to preserve the <empty>-not-propagating. I propose that if the var is expanded without being quoted, then setprop will syntax error out under that condition, whereas if they were quoted, then the property would be set empty.

Also note, in internal master (and likely AOSP), somehow the error logging for /init is being attributed to ext4_utils. That is yet another issue ...

Patch-set: 2