Update patch set 2
Patch Set 2:

Problem is when one specifies many cases, and then =*, they _will_ get racy and unexpected behaviors. =* is IMHO _useless_ in combination with the others. Our current usage pattern is to treat =* as orthogonal to any specific cases, use one of the others, not both.

An alternative is to allow _not_  equal (!=) for a property condition and then use it like:

on property:<property>!= && property:<property>=*

Patch-set: 2