Update patch set 2
Patch Set 2:

My worry is if multiple users trigger off of the same property, e.g. user A triggers off of specific properties and user B triggers off of =* and that causes a clash.  I don't see this playing out in the files that we currently have though.  

!= is a possibility.

We support any number of property triggers chained together.

Patch-set: 2