Update patch set 6
Patch Set 6:

This should reflect most comments.

I left a different README.txt, modeled on the bionic one.  The bionic one was long and different enough that short instructions seemed appropriate here.

I removed the RefBase.cpp changes, deferring some variant of those to a future CL.  I left enough stuff in this CL to actually make it possible to run the tests and have them pass. That seemed to sufficiently central to the point of this CL.

Patch-set: 6
Label: Verified=0