Update patch set 1
Patch Set 1:

Specifically for non-QC specific compositions (MTP, MTP+ADB, PTP, PTP+ADB, etc) we only have the rules defined in init.usb.configfs.rc, and do not have anything further in init.$platform.usb.rc. @Badhri are you saying that we could/should further have our own rules for the same composition triggers in init.$platform.usb.rc so that we can fill in VID/PID there? Maybe this is misunderstanding on our part, but how do triggers get invoked if they are defined in multiple RC files?

Also note we have never defined our own specific PIDs for the standard Android compositions which is why we are fine using the Google IDs for those, hence we thought it would be appropriate to make the change in the common init.usb.configfs.rc (to follow the case for init.usb.rc for non-ConfigFS)

Patch-set: 1