Update patch set 4
Patch Set 4:

(6 comments)

I think I found all of them, but by now I'm not sure.
Mark, it's up to you whether you want to fix those or not; strictly speaking these are buffer access violations, but apparently this issue existed for a while so it could be a lower priority to address those.

Patch-set: 4
Label: Verified=0