Update patch set 2
Patch Set 2:

I made the assumption that if the file is inaccessible, blocked, or missing that we preserve the logpersist data in order to help diagnose; and also allow an API to _ignore_ the signature change by chmod 000 to the file to prevent the loss of data.

Thanks, I forgot to add a test for that in logcat_test.cpp.

Patch-set: 2